MCBSTM32F400: Refactor Initialization and TinyHal libraries
