--- Solutions/MCBSTM32F400/DeviceCode/TinyHal/MasterConfig.h
@@ -1,23 +1,0 @@
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Copyright (c) Microsoft Corporation.  All rights reserved.
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-//--//
-
-#ifndef _MASTER_CONFIG_H_
-#define _MASTER_CONFIG_H_ 1
-
-//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//
-
-#include <platform_selector.h>
-
-//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//
-
-#define NUM_DEBUGGERS          1
-#define NUM_MESSAGING          1
-
-#define DEBUGGER_PORT_INDEX    0
-
-//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//--//
-
-#endif  // _MASTER_CONFIG_H_
