+++ Solutions/MCBSTM32F400/DeviceCode/TinyHal/dotNetMF.proj
@@ -5,25 +5,25 @@
         <ProjectGuid>{b65a0478-f1a0-417a-9c5e-88d4424a1696}</ProjectGuid>
         <Size>
         </Size>
-        <Description>System initialization library</Description>
+        <Description>MCBSTM32F400 tinyhal library</Description>
         <Level>HAL</Level>
         <LibraryFile>MCBSTM32F400_tinyhal.$(LIB_EXT)</LibraryFile>
         <ProjectPath>$(SPOCLIENT)\Solutions\MCBSTM32F400\DeviceCode\TinyHal\dotNetMF.proj</ProjectPath>
         <ManifestFile>MCBSTM32F400_tinyhal.$(LIB_EXT).manifest</ManifestFile>
         <Groups>System</Groups>
-        <LibraryCategory>
-            <MFComponent xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" Name="Initialization_HAL" Guid="{34EDFFBE-0FB6-49CC-A673-03BD96EDE0A9}" ProjectPath="" Conditional="" xmlns="">
-                <VersionDependency xmlns="http://schemas.microsoft.com/netmf/InventoryFormat.xsd">
-                    <Major>4</Major>
-                    <Minor>0</Minor>
-                    <Revision>0</Revision>
-                    <Build>0</Build>
-                    <Extra />
-                    <Date>2009-04-30</Date>
-                </VersionDependency>
-                <ComponentType xmlns="http://schemas.microsoft.com/netmf/InventoryFormat.xsd">LibraryCategory</ComponentType>
-            </MFComponent>
-        </LibraryCategory>
+		<LibraryCategory>
+		  <MFComponent xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" Name="TinyHal_HAL" Guid="{37C37BC0-3CA6-4C1C-A26F-4761A7BA3C05}" ProjectPath="" Conditional="" xmlns="">
+			<VersionDependency xmlns="http://schemas.microsoft.com/netmf/InventoryFormat.xsd">
+			  <Major>4</Major>
+			  <Minor>0</Minor>
+			  <Revision>0</Revision>
+			  <Build>0</Build>
+			  <Extra />
+			  <Date>2009-04-30</Date>
+			</VersionDependency>
+			<ComponentType xmlns="http://schemas.microsoft.com/netmf/InventoryFormat.xsd">LibraryCategory</ComponentType>
+		  </MFComponent>
+		</LibraryCategory>
         <Documentation>
         </Documentation>
         <PlatformIndependent>False</PlatformIndependent>
@@ -36,6 +36,7 @@
         <PlatformIndependentBuild>false</PlatformIndependentBuild>
         <Version>4.0.0.0</Version>
     </PropertyGroup>
+    <Import Project="$(SPOCLIENT)\tools\targets\Microsoft.SPOT.System.Settings" />
     <Import Project="$(SPOCLIENT)\Framework\Features\BlockStorage_PAL.libcatproj" />
     <Import Project="$(SPOCLIENT)\Framework\Features\BlockStorageAddDevices_HAL.libcatproj" />
     <Import Project="$(SPOCLIENT)\Framework\Features\Buttons_PAL.libcatproj" />
@@ -47,7 +48,6 @@
     <Import Project="$(SPOCLIENT)\Framework\Features\Time_PAL.libcatproj" />
     <Import Project="$(SPOCLIENT)\Framework\Features\TimeService_PAL.libcatproj" />
     <Import Project="$(SPOCLIENT)\Framework\Features\Heap_PAL.libcatproj" />
-    <Import Project="$(SPOCLIENT)\tools\targets\Microsoft.SPOT.System.Settings" />
     <PropertyGroup>
         <OutputType>Library</OutputType>
     </PropertyGroup>
@@ -59,10 +59,7 @@
         <Compile Include="$(AS_SUBDIR)\Interlocked.s" />
     </ItemGroup>
     <ItemGroup>
-        <Compile Include="OEM_Model_SKU_NetworkID.cpp" />
-        <Compile Include="tinyhal.cpp" />
         <Compile Include="Aborts.cpp" />
-        <Compile Include="SystemState.cpp" />
         <Compile Include="CpuPower.cpp"/>
         <Compile Include="Retarget.cpp"/>
         <Compile Include="main.cpp" />
